feat: add dockerfile
Adds our dockerfile based on mkdocs and our custom core plugin.

In contrast to the Backstage techdocs-container [0], we have decided not
to base our image on alpine, but bullseye-slim instead. As we have
removed the PlantUML dependency and opted for using kroki [1] for
flowchart rendering in our core plugin, we managed to reduce the image
size significantly, thus eliminating the need to use image size
reduction techniques. As an additional benefit this allows us to keep
the dockerfile short and sweet and increase build times.

[0] https://github.com/backstage/techdocs-container
[1] https://kroki.io

Signed-off-by: Viktor Bahr <viktor.bahr@tier.app>